major cleanup + loads of fixes and restructuring
