+++ hyperiax/execution/fastexecutor.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 from abc import ABC
-from ..tree.fasttree import FastTree
+from ..tree import HypTree
 import jax
-from ..models import UpModel, UpDownModel, DownModel, FuseModel
+from ..models import UpModel, DownModel, FuseModel
 from inspect import getfullargspec
 from functools import partial
 
@@ -15,55 +15,32 @@ class OrderedExecutor(ABC):
     are to be executed in the tree. It is assumed that all operations are batched.
     """
     def __init__(self, 
-                 model : UpModel | UpDownModel | DownModel,
+                 model : UpModel | DownModel,
                  key = None) -> None:
         self.key = key if key else jax.random.PRNGKey(0)
         self.model = model
 
-        if issubclass(type(self.model), (UpDownModel, DownModel)):
-            self._set_down_keys()
-        if issubclass(type(self.model), (UpDownModel, UpModel)):
-            self._set_up_keys()
-        if issubclass(type(self.model), FuseModel):
-            self._set_fuse_keys()
-
-    def updown(self, tree, params = {}):
-        """Runs both the up and downward pass
-
-        Args:
-            tree (HypTree): The tree to execute on
-            params (dict, optional): optional parameters for mcmc. Defaults to {}.
-
-        Returns:
-            HypTree: The tree after running one up and down pass
-        """
-        t = self.up(tree, params=params)
-        t = self.down(t, params=params)
-
-        return t
-    
-
-    def up(self, tree : FastTree, params = {}):
-        if not issubclass(type(self.model), (UpDownModel, UpModel)):
-            raise ValueError('Model needs to be of type UpDownModel or UpModel')
+    def up(self, tree : HypTree, params = {}):
+        if not issubclass(type(self.model), UpModel):
+            raise ValueError('Model needs to be of type UpModel')
         # out here we fetch the data stored in the tree
 
         #for k,val_shape in self.model.up_produces.items():
         #    if k not in tree.data:
         #        tree.add_property(k, val_shape)
 
-        if issubclass(type(self.model), (UpDownModel, UpModel)):
+        if issubclass(type(self.model), UpModel):
             new_data = self._up_inner(tree.data, tree, params)
 
             tree.data = {**tree.data, **new_data}
 
-            return new_data
+            return 
         elif issubclass(type(self.model), FuseModel):
             new_data = self._fuse(tree, params = params)
 
             tree.data = {**tree.data, **new_data}
 
-            return new_data
+            return
 
     @partial(jax.jit, static_argnames=['tree', 'self']) 
     def _up_inner(self, data, tree, params):
@@ -75,7 +52,7 @@ class OrderedExecutor(ABC):
         for (level_start, level_end), up_ref in iterator:
             node_data = {
                 k: jax.lax.slice_in_dim(data[k], level_start, level_end)
-                for k in self.up_keys
+                for k in self.model.up_keys
             }
             #print(node_data)
             up_result = self.model.up(**node_data, params = params)
@@ -89,7 +66,7 @@ class OrderedExecutor(ABC):
 
             parent_data = {
                 k: data[k][up_ref]
-                for k in self.transform_parent_keys
+                for k in self.model.transform_parent_keys
             }
 
             result = self.model.transform(**parent_data, **fuse_scatter, params = params)
@@ -99,44 +76,7 @@ class OrderedExecutor(ABC):
 
         return data
 
-            
-
-
-    def _set_up_keys(self):
-        up_arg_spec = getfullargspec(self.model.up)
-        upkeys = up_arg_spec.args
-        self.up_keys = upkeys
-
-        transform_arg_spec = getfullargspec(self.model.transform)
-        keys = transform_arg_spec.args
-        transform_child_keys = [k.removeprefix('child_') for k in keys if k.startswith('child_')]
-        transform_parent_keys = [k for k in keys if not k.startswith('child_')]
-
-        self.transform_child_keys = transform_child_keys
-        self.transform_parent_keys = transform_parent_keys
-
-
-    def _set_fuse_keys(self): 
-        fuse_arg_spec = getfullargspec(self.model.transform)
-        keys = fuse_arg_spec.args
-        fuse_child_keys = [k.removeprefix('child_') for k in keys if k.startswith('child_')]
-        fuse_parent_keys = [k for k in keys if not k.startswith('child_')]
-
-        self.fuse_child_keys = fuse_child_keys
-        self.fuse_parent_keys = fuse_parent_keys
-
-    def _set_down_keys(self):
-        arg_spec = getfullargspec(self.model.down)
-
-        keys = arg_spec.args # TODO: we should remove things like key, up_msg, params
-
-        down_parent_keys = [k.removeprefix('parent_') for k in keys if k.startswith('parent_')]
-        down_child_keys = [k for k in keys if not k.startswith('parent_')]
-
-        self.down_parent_keys = down_parent_keys
-        self.down_child_keys = down_child_keys
-
-    def _fuse(self, tree : FastTree, params = {}):
+    def _fuse(self, tree : HypTree, params = {}):
         if not issubclass(type(self.model), FuseModel):
             raise ValueError('Model needs to be of type FuseModel')
         # out here we fetch the data stored in the tree
@@ -153,12 +93,12 @@ class OrderedExecutor(ABC):
         for level_start, level_end in tree.levels[1:]:
             node_data = {
                 k: jax.lax.slice_in_dim(data[k], level_start, level_end)
-                for k in self.down_child_keys
+                for k in self.model.down_child_keys
             }
             parent_indices = tree.parents[level_start:level_end] # no need to lax since tree is untraced
             parent_data = {
                 f'parent_{k}': data[k][parent_indices]
-                for k in self.down_parent_keys
+                for k in self.model.down_parent_keys
             }
 
             result = self.model.down(**node_data, **parent_data)
@@ -167,7 +107,7 @@ class OrderedExecutor(ABC):
                 data[k] = data[k].at[level_start:level_end].set(val)
         return data
     
-    def down(self, tree : FastTree, params = {}):
+    def down(self, tree : HypTree, params = {}):
         """Runs the down pass on the tree using the `down` function from the model.
 
         The `down` function in the model takes the following special parameters;
@@ -191,8 +131,8 @@ class OrderedExecutor(ABC):
         Returns:
             HypTree: The tree after running one pass
         """
-        if not issubclass(type(self.model), (UpDownModel, DownModel)):
-            raise ValueError('Model needs to be of type UpDownModel or DownModel')
+        if not issubclass(type(self.model), DownModel):
+            raise ValueError('Model needs to be of type DownModel')
         # out here we fetch the data stored in the tree
 
         new_data = self._down_inner(tree.data, tree)
@@ -200,6 +140,3 @@ class OrderedExecutor(ABC):
         tree.data = {**tree.data, **new_data}
 
         return 
-    
-    def __call__(self, tree, params = {}):
-        return self.updown(tree, params = params)
