+++ hyperiax/models/__init__.py
@@ -1,4 +1,4 @@
-from .updownmodel import UpModel,DownModel,UpDownModel,FuseModel
-from .lambdamodels import UpLambda, DownLambda, UpDownLambda, FuseLambda, UpdateLambda
+from .updownmodel import UpModel,DownModel,FuseModel
+from .lambdamodels import UpLambda, DownLambda, FuseLambda, UpdateLambda
 from .updatemodel import UpdateModel
 from . import functional
