+++ hyperiax/models/lambdamodels.py
@@ -1,4 +1,4 @@
-from .updownmodel import UpModel, UpDownModel, DownModel, FuseModel
+from .updownmodel import UpModel, DownModel, FuseModel
 from .updatemodel import UpdateModel
 
 class UpLambda(UpModel):
@@ -12,10 +12,11 @@ class UpLambda(UpModel):
         :param up_fn: param intputs from node to transform_fn
         :param transform_fn: function to transform or do calculations form up parameters
         """
-        super().__init__(reductions=reductions)
 
         self.up = up_fn
         self.transform = transform_fn
+
+        super().__init__(reductions=reductions)
 
     def up(self, *args, **kwargs):
         """ Up function to define values to fuse function
@@ -29,36 +30,6 @@ class UpLambda(UpModel):
         """
         raise ValueError('Model does not have a valid fuse function')
     
-class UpDownLambda(UpDownModel):
-    """Lambda model that only contains both an up and down interface.  
-
-    :param UpDownModel: Requires an up, fuse and down function.
-    """
-    def __init__(self, up_fn, fuse_fn, down_fn) -> None:
-        super().__init__()
-
-        self.up_fn = up_fn
-        self.fuse_fn = fuse_fn
-        self.down_fn = down_fn
-
-    def up(self, *args, **kwargs):
-        """ Up function to define values to fuse function
-        :return: input arguments to fuse function
-        """
-        return self.up_fn(*args, **kwargs)
-    
-    def fuse(self, *args, **kwargs):
-        """ Fuse function to define calculations from up parameters
-        :return: calculated values from up parameters to parent node
-        """
-        return self.fuse_fn(*args, **kwargs)
-    
-    def down(self, *args, **kwargs):
-        """ Down function to define values to fuse function
-        :return: input arguments to fuse function
-        """
-        return self.down_fn(*args, **kwargs)
-    
 class DownLambda(DownModel):
     """ Lambda model that only contains a down interface
 
@@ -70,10 +41,10 @@ class DownLambda(DownModel):
 
         :param down_fn: function to define values to fuse function
         """
-        super().__init__()
-
         self.down = down_fn # need to do this to carry the argspec
     
+        super().__init__()
+
     def down(self, *args, **kwargs):
         """ Down function to define values to fuse function
         :return: input arguments to fuse function
@@ -91,10 +62,10 @@ class UpdateLambda(UpdateModel):
         :param up_fn: param intputs from node to update_fn
         :param update_fn: function to update or do calculations form up parameters
         """
-        super().__init__(reductions=reductions)
-
         self.up = up_fn
         self.update = update_fn
+
+        super().__init__(reductions=reductions)
 
     def up(self, *args, **kwargs):
         """ Up function to define values to fuse function
@@ -118,9 +89,9 @@ class FuseLambda(FuseModel):
 
         :param fuse_fn: function to define values to fuse function
         """
-        super().__init__()
 
         self.fuse = fuse_fn
+        super().__init__()
 
     def fuse(self, *args, **kwargs):
         """ Fuse function to define values to fuse function
