+++ hyperiax/plotting/__init__.py
@@ -1,1 +1,2 @@
 from .scatter import plot_tree_2d_scatter
+from .ascii import plot_tree_text
