+++ hyperiax/tree/__init__.py
@@ -1,2 +1,1 @@
-from .tree import HypTree, TreeNode
-from . import builders, initializers
+from .tree import HypTree, TopologyNode
